Move calibration back to ParameterTypeProcessor
While splitting the calibration logic out of
ParameterTypeProcessor, it was put in a separate
class, but in retrospect this wasn't really
necessary. Move it back to ParameterTypeProcessor
for old times sake.
